micro cleanup
