+++ .github/CODEOWNERS
@@ -1,4 +1,4 @@
 * @atomiks @michaldudak
-/docs/data/ @atomiks @colmtuite @michaldudak
+/docs/src/ @atomiks @colmtuite @michaldudak
 /packages/react/ @atomiks @colmtuite @michaldudak
 /scripts/ @michaldudak
