+++ docs/src/components/icons/GitHubIcon.tsx
