+++ docs/src/components/icons/NpmIcon.tsx
